Cleaned API code
