+++ src/javascript/Api/ContentEditorApi.scss
@@ -1,0 +1,25 @@
+.ceDialogRoot.ceDialogRoot {
+    right: 0;
+    left: unset;
+
+    // Reduce zIndex to be able to display the old edit engine tabs
+    z-index: 1031;
+
+    width: calc(100vw - 56px);
+}
+
+.ceWindowRoot.ceWindowRoot {
+    position: absolute;
+    right: 0;
+    bottom: 0;
+    z-index: 1500;
+
+    display: flex;
+    flex-direction: column;
+    width: 40vw;
+    min-width: 620px;
+    height: calc(100vh - 124px);
+
+    box-shadow: 0 4px 8px var(--color-dark40);
+    opacity: 1;
+}
