+++ src/javascript/Api/ContentTypeSelectorModal.jsx
@@ -1,0 +1,47 @@
+import {CreateNewContentDialog} from '~/Create/CreateNewContentAction/CreateNewContentDialog';
+import {Constants} from '~/ContentEditor.constants';
+import React from 'react';
+import PropTypes from 'prop-types';
+
+export const ContentTypeSelectorModal = ({
+    contentTypeSelectorConfig,
+    setContentTypeSelectorConfig,
+    setEditorConfig
+}) => {
+    return (
+        <CreateNewContentDialog
+            open
+            childNodeName={contentTypeSelectorConfig.name}
+            nodeTypes={contentTypeSelectorConfig.creatableNodeTypes}
+            includeSubTypes={contentTypeSelectorConfig.includeSubTypes}
+            parentPath={contentTypeSelectorConfig.path}
+            uilang={contentTypeSelectorConfig.uilang}
+            onClose={() => {
+                setContentTypeSelectorConfig(false);
+            }}
+            onExited={() => {
+                setContentTypeSelectorConfig(false);
+            }}
+            onCreateContent={contentType => {
+                setContentTypeSelectorConfig(false);
+                setEditorConfig({
+                    name: contentTypeSelectorConfig.name,
+                    uuid: contentTypeSelectorConfig.uuid,
+                    site: contentTypeSelectorConfig.site,
+                    uilang: contentTypeSelectorConfig.uilang,
+                    initLang: contentTypeSelectorConfig.lang,
+                    lang: contentTypeSelectorConfig.lang,
+                    isFullscreen: contentTypeSelectorConfig.isFullscreen,
+                    contentType: contentType.name,
+                    mode: Constants.routes.baseCreateRoute
+                });
+            }}
+        />
+    );
+};
+
+ContentTypeSelectorModal.propTypes = {
+    contentTypeSelectorConfig: PropTypes.object.isRequired,
+    setContentTypeSelectorConfig: PropTypes.func.isRequired,
+    setEditorConfig: PropTypes.func.isRequired
+};
