+++ src/javascript/Edit/EditContent.action.jsx
@@ -6,7 +6,7 @@ import {useSelector} from 'react-redux';
 import {Constants} from '~/ContentEditor.constants';
 import {useTranslation} from 'react-i18next';
 
-export const EditContent = ({path, isWindow, editCallback, render: Render, loading: Loading, ...otherProps}) => {
+export const EditContent = ({path, isModal, isFullscreen, editCallback, render: Render, loading: Loading, ...otherProps}) => {
     const {redirect} = useContentEditorHistory();
     useTranslation('content-editor');
     const {language, uilang, site} = useSelector(state => ({language: state.language, site: state.site, uilang: state.uilang}));
@@ -22,20 +22,22 @@ export const EditContent = ({path, isWindow, editCallback, render: Render, loadi
     return (
         <Render {...otherProps}
                 isVisible={res.checksResult}
-                onClick={() => isWindow ? window.CE_API.edit(res.node.uuid, site, language, uilang, true, editCallback) : redirect({language, mode: Constants.routes.baseEditRoute, uuid: res.node.uuid})}
+                onClick={() => isModal ? window.CE_API.edit(res.node.uuid, site, language, uilang, isFullscreen, editCallback) : redirect({language, mode: Constants.routes.baseEditRoute, uuid: res.node.uuid})}
         />
     );
 };
 
 EditContent.defaultProps = {
     loading: undefined,
-    isWindow: false,
+    isModal: false,
+    isFullscreen: false,
     editCallback: undefined
 };
 
 EditContent.propTypes = {
     path: PropTypes.string.isRequired,
-    isWindow: PropTypes.bool,
+    isModal: PropTypes.bool,
+    isFullscreen: PropTypes.bool,
     editCallback: PropTypes.func,
     render: PropTypes.func.isRequired,
     loading: PropTypes.func
